move pilosa/test/client.go helpers into pilosa/client_test.go
